Refactor MCP tools: consolidate color definitions into HexStrikeColors class and register core network scanning tools
