+++ hexstrike_mcp.py
@@ -21,68 +21,7 @@ import core.config_core as config_core
 
 from mcp.server.fastmcp import FastMCP
 
-class HexStrikeColors:
-    """Enhanced color palette matching the server's ModernVisualEngine.COLORS"""
-
-    # Basic colors (for backward compatibility)
-    RED = '\033[91m'
-    GREEN = '\033[92m'
-    YELLOW = '\033[93m'
-    BLUE = '\033[94m'
-    MAGENTA = '\033[95m'
-    CYAN = '\033[96m'
-    WHITE = '\033[97m'
-
-    # Core enhanced colors
-    MATRIX_GREEN = '\033[38;5;46m'
-    NEON_BLUE = '\033[38;5;51m'
-    ELECTRIC_PURPLE = '\033[38;5;129m'
-    CYBER_ORANGE = '\033[38;5;208m'
-    HACKER_RED = '\033[38;5;196m'
-    TERMINAL_GRAY = '\033[38;5;240m'
-    BRIGHT_WHITE = '\033[97m'
-    RESET = '\033[0m'
-    BOLD = '\033[1m'
-    DIM = '\033[2m'
-
-    # Enhanced reddish tones and highlighting colors
-    BLOOD_RED = '\033[38;5;124m'
-    CRIMSON = '\033[38;5;160m'
-    DARK_RED = '\033[38;5;88m'
-    FIRE_RED = '\033[38;5;202m'
-    ROSE_RED = '\033[38;5;167m'
-    BURGUNDY = '\033[38;5;52m'
-    SCARLET = '\033[38;5;197m'
-    RUBY = '\033[38;5;161m'
-
-    # Highlighting colors
-    HIGHLIGHT_RED = '\033[48;5;196m\033[38;5;15m'  # Red background, white text
-    HIGHLIGHT_YELLOW = '\033[48;5;226m\033[38;5;16m'  # Yellow background, black text
-    HIGHLIGHT_GREEN = '\033[48;5;46m\033[38;5;16m'  # Green background, black text
-    HIGHLIGHT_BLUE = '\033[48;5;51m\033[38;5;16m'  # Blue background, black text
-    HIGHLIGHT_PURPLE = '\033[48;5;129m\033[38;5;15m'  # Purple background, white text
-
-    # Status colors with reddish tones
-    SUCCESS = '\033[38;5;46m'  # Bright green
-    WARNING = '\033[38;5;208m'  # Orange
-    ERROR = '\033[38;5;196m'  # Bright red
-    CRITICAL = '\033[48;5;196m\033[38;5;15m\033[1m'  # Red background, white bold text
-    INFO = '\033[38;5;51m'  # Cyan
-    DEBUG = '\033[38;5;240m'  # Gray
-
-    # Vulnerability severity colors
-    VULN_CRITICAL = '\033[48;5;124m\033[38;5;15m\033[1m'  # Dark red background
-    VULN_HIGH = '\033[38;5;196m\033[1m'  # Bright red bold
-    VULN_MEDIUM = '\033[38;5;208m\033[1m'  # Orange bold
-    VULN_LOW = '\033[38;5;226m'  # Yellow
-    VULN_INFO = '\033[38;5;51m'  # Cyan
-
-    # Tool status colors
-    TOOL_RUNNING = '\033[38;5;46m\033[5m'  # Blinking green
-    TOOL_SUCCESS = '\033[38;5;46m\033[1m'  # Bold green
-    TOOL_FAILED = '\033[38;5;196m\033[1m'  # Bold red
-    TOOL_TIMEOUT = '\033[38;5;208m\033[1m'  # Bold orange
-    TOOL_RECOVERY = '\033[38;5;129m\033[1m'  # Bold purple
+from mcp_core.hexstrikecolors import HexStrikeColors
 
 # Backward compatibility alias
 Colors = HexStrikeColors
@@ -343,162 +282,10 @@ def setup_mcp_server(hexstrike_client: HexStrikeClient, compact: bool = False) -
     from mcp_tools.database import register_database_tools
     register_database_tools(mcp, hexstrike_client, logger)
 
-    # ============================================================================
-    # CORE NETWORK SCANNING TOOLS
-    # ============================================================================
-
-    @mcp.tool()
-    def whois_lookup(target: str) -> Dict[str, Any]:
-        """
-        Perform a WHOIS lookup for a domain or IP address.
-
-        Args:
-            target: The domain or IP to query
-
-        Returns:
-            WHOIS lookup results
-        """
-        try:
-            response = hexstrike_client.safe_post("api/tools/whois", {"target": target})
-            return response
-        except Exception as e:
-            logger.error(f"WHOIS lookup failed: {e}")
-            return {"error": str(e)}
+    # Register core network scanning tools
+    from mcp_tools.core_network_scanning import register_core_network_scanning_tools
+    register_core_network_scanning_tools(mcp, hexstrike_client, logger, HexStrikeColors)
     
-    @mcp.tool()
-    def nmap_scan(target: str, scan_type: str = "-sV", ports: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute an enhanced Nmap scan against a target with real-time logging.
-
-        Args:
-            target: The IP address or hostname to scan
-            scan_type: Scan type (e.g., -sV for version detection, -sC for scripts)
-            ports: Comma-separated list of ports or port ranges
-            additional_args: Additional Nmap arguments
-
-        Returns:
-            Scan results with enhanced telemetry
-        """
-        data: Dict[str, Any] = {
-            "target": target,
-            "scan_type": scan_type,
-            "ports": ports,
-            "additional_args": additional_args
-        }
-        logger.info(f"{HexStrikeColors.FIRE_RED}🔍 Initiating Nmap scan: {target}{HexStrikeColors.RESET}")
-
-        # Use enhanced error handling by default
-        data["use_recovery"] = True
-        result = hexstrike_client.safe_post("api/tools/nmap", data)
-
-        if result.get("success"):
-            logger.info(f"{HexStrikeColors.SUCCESS}✅ Nmap scan completed successfully for {target}{HexStrikeColors.RESET}")
-
-            # Check for recovery information
-            if result.get("recovery_info", {}).get("recovery_applied"):
-                recovery_info = result["recovery_info"]
-                attempts = recovery_info.get("attempts_made", 1)
-                logger.info(f"{HexStrikeColors.HIGHLIGHT_YELLOW} Recovery applied: {attempts} attempts made {HexStrikeColors.RESET}")
-        else:
-            logger.error(f"{HexStrikeColors.ERROR}❌ Nmap scan failed for {target}{HexStrikeColors.RESET}")
-
-            # Check for human escalation
-            if result.get("human_escalation"):
-                logger.error(f"{HexStrikeColors.CRITICAL} HUMAN ESCALATION REQUIRED {HexStrikeColors.RESET}")
-
-        return result
-
-    @mcp.tool()
-    def gobuster_scan(url: str, mode: str = "dir", wordlist: str = "/usr/share/wordlists/dirb/common.txt", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Gobuster to find directories, DNS subdomains, or virtual hosts with enhanced logging.
-
-        Args:
-            url: The target URL
-            mode: Scan mode (dir, dns, fuzz, vhost)
-            wordlist: Path to wordlist file
-            additional_args: Additional Gobuster arguments
-
-        Returns:
-            Scan results with enhanced telemetry
-        """
-        data: Dict[str, Any] = {
-            "url": url,
-            "mode": mode,
-            "wordlist": wordlist,
-            "additional_args": additional_args
-        }
-        logger.info(f"{HexStrikeColors.CRIMSON}📁 Starting Gobuster {mode} scan: {url}{HexStrikeColors.RESET}")
-
-        # Use enhanced error handling by default
-        data["use_recovery"] = True
-        result = hexstrike_client.safe_post("api/tools/gobuster", data)
-
-        if result.get("success"):
-            logger.info(f"{HexStrikeColors.SUCCESS}✅ Gobuster scan completed for {url}{HexStrikeColors.RESET}")
-
-            # Check for recovery information
-            if result.get("recovery_info", {}).get("recovery_applied"):
-                recovery_info = result["recovery_info"]
-                attempts = recovery_info.get("attempts_made", 1)
-                logger.info(f"{HexStrikeColors.HIGHLIGHT_YELLOW} Recovery applied: {attempts} attempts made {HexStrikeColors.RESET}")
-        else:
-            logger.error(f"{HexStrikeColors.ERROR}❌ Gobuster scan failed for {url}{HexStrikeColors.RESET}")
-
-            # Check for alternative tool suggestion
-            if result.get("alternative_tool_suggested"):
-                alt_tool = result["alternative_tool_suggested"]
-                logger.info(f"{HexStrikeColors.HIGHLIGHT_BLUE} Alternative tool suggested: {alt_tool} {HexStrikeColors.RESET}")
-
-        return result
-
-    @mcp.tool()
-    def nuclei_scan(target: str, severity: str = "", tags: str = "", template: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Nuclei vulnerability scanner with enhanced logging and real-time progress.
-
-        Args:
-            target: The target URL or IP
-            severity: Filter by severity (critical,high,medium,low,info)
-            tags: Filter by tags (e.g. cve,rce,lfi)
-            template: Custom template path
-            additional_args: Additional Nuclei arguments
-
-        Returns:
-            Scan results with discovered vulnerabilities and telemetry
-        """
-        data: Dict[str, Any] = {
-            "target": target,
-            "severity": severity,
-            "tags": tags,
-            "template": template,
-            "additional_args": additional_args
-        }
-        logger.info(f"{HexStrikeColors.BLOOD_RED}🔬 Starting Nuclei vulnerability scan: {target}{HexStrikeColors.RESET}")
-
-        # Use enhanced error handling by default
-        data["use_recovery"] = True
-        result = hexstrike_client.safe_post("api/tools/nuclei", data)
-
-        if result.get("success"):
-            logger.info(f"{HexStrikeColors.SUCCESS}✅ Nuclei scan completed for {target}{HexStrikeColors.RESET}")
-
-            # Enhanced vulnerability reporting
-            if result.get("stdout") and "CRITICAL" in result["stdout"]:
-                logger.warning(f"{HexStrikeColors.CRITICAL} CRITICAL vulnerabilities detected! {HexStrikeColors.RESET}")
-            elif result.get("stdout") and "HIGH" in result["stdout"]:
-                logger.warning(f"{HexStrikeColors.FIRE_RED} HIGH severity vulnerabilities found! {HexStrikeColors.RESET}")
-
-            # Check for recovery information
-            if result.get("recovery_info", {}).get("recovery_applied"):
-                recovery_info = result["recovery_info"]
-                attempts = recovery_info.get("attempts_made", 1)
-                logger.info(f"{HexStrikeColors.HIGHLIGHT_YELLOW} Recovery applied: {attempts} attempts made {HexStrikeColors.RESET}")
-        else:
-            logger.error(f"{HexStrikeColors.ERROR}❌ Nuclei scan failed for {target}{HexStrikeColors.RESET}")
-
-        return result
-
     # ============================================================================
     # CLOUD SECURITY TOOLS
     # ============================================================================
